refactor(web/stores): extract useLyricsAppearanceStore from useAppStore
Move the four lyrics appearance fields (plain/synced opacity + font
size), their setters, the resetLyricsPlain/resetLyricsAppearance
helpers, the four coerce/clamp validators, and the LyricsFontSize /
LYR_SIZE_CLASS / nextLyricsFontSize / LYRICS_*_PAST_RATIO constants
into a dedicated useLyricsAppearanceStore.

The new store persists to its own bucket (shiranami.lyrics-appearance-
store) and runs a one-shot import-time migration that lifts existing
lyrics state out of the pre-split combined shiranami.app-store key on
first load. The legacy bucket is left intact so its other (UI/compact)
fields remain reachable for their own future migrations. Persistence
shape on disk for users with no existing lyrics state in the legacy
bucket is unchanged.

Direct consumers (LyricsPanel, LyricsSection, NowPlayingView) now read
from useLyricsAppearanceStore. useAppStore re-exports the lyrics
constants/types so the remaining indirect imports keep compiling
through this PR; those will follow when the rest of the split lands.
The lyrics-specific tests move into useLyricsAppearanceStore.test.ts
along with three legacy-bucket sanitization cases.
