+++ apps/web/src/components/now-playing/NowPlayingView.tsx
@@ -2,8 +2,12 @@ import { useEffect, useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
 import { toast } from 'sonner';
 import { usePlaybackStore } from '@/stores/usePlaybackStore';
-import type { LyricsFontSize } from '@/stores/useAppStore';
-import { useAppStore, LYRICS_SYNCED_PAST_RATIO } from '@/stores/useAppStore';
+import { useAppStore } from '@/stores/useAppStore';
+import {
+  useLyricsAppearanceStore,
+  LYRICS_SYNCED_PAST_RATIO,
+  type LyricsFontSize,
+} from '@/stores/useLyricsAppearanceStore';
 import { useViewStore } from '@/stores/useViewStore';
 import { useLyricsQuery } from '@/hooks/queries/useLyrics';
 import { useActiveLineIndex } from '@/lib/lyrics';
@@ -55,10 +59,10 @@ export function NowPlayingView() {
   const exitNowPlaying = useViewStore(s => s.exitNowPlaying);
   const lyricsVisible = useAppStore(s => s.nowPlayingLyricsVisible);
   const toggleLyrics = useAppStore(s => s.toggleNowPlayingLyrics);
-  const lyricsPlainOpacity = useAppStore(s => s.lyricsPlainOpacity);
-  const lyricsPlainFontSize = useAppStore(s => s.lyricsPlainFontSize);
-  const lyricsSyncedDimOpacity = useAppStore(s => s.lyricsSyncedDimOpacity);
-  const lyricsSyncedFontSize = useAppStore(s => s.lyricsSyncedFontSize);
+  const lyricsPlainOpacity = useLyricsAppearanceStore(s => s.lyricsPlainOpacity);
+  const lyricsPlainFontSize = useLyricsAppearanceStore(s => s.lyricsPlainFontSize);
+  const lyricsSyncedDimOpacity = useLyricsAppearanceStore(s => s.lyricsSyncedDimOpacity);
+  const lyricsSyncedFontSize = useLyricsAppearanceStore(s => s.lyricsSyncedFontSize);
 
   const npBase = cn(NP_BASE_SHARED, NP_SYNCED_BASE_SIZE_CLASS[lyricsSyncedFontSize]);
   const npActive = cn(
