+++ apps/web/src/stores/useAppStore.ts
@@ -2,16 +2,16 @@ import { create } from 'zustand';
 import { persist, createJSONStorage } from 'zustand/middleware';
 import { IS_ELECTRON } from '@/lib/platform';
 import { useViewStore, type AppView, type RightPanel } from '@/stores/useViewStore';
+import type { LyricsFontSize } from '@/stores/useLyricsAppearanceStore';
 
-export type { AppView, RightPanel };
+export type { AppView, RightPanel, LyricsFontSize };
+/** @deprecated Use LyricsFontSize. Alias kept for back-compat with existing imports. */
+export type LyricsPlainFontSize = LyricsFontSize;
 export type VisualizerStyle = 'bars' | 'waveform' | 'circle' | 'particles';
 export type LibraryViewMode = 'tracks' | 'albums';
 export type AlbumGridSize = 'small' | 'medium' | 'large';
 export type AlbumSortMode = 'name' | 'artist' | 'year' | 'recentlyAdded';
 export type AlbumSortOrder = 'asc' | 'desc';
-export type LyricsFontSize = 'sm' | 'base' | 'lg' | 'xl';
-/** @deprecated Use LyricsFontSize. Alias kept for back-compat with existing imports. */
-export type LyricsPlainFontSize = LyricsFontSize;
 
 export type CompactSize = 'sm' | 'md' | 'lg';
 export type CompactFontSize = 'sm' | 'md' | 'lg';
@@ -22,31 +22,21 @@ export const UI_SCALE_DEFAULT = 100;
 export const UI_SCALE_STEP = 5;
 export const UI_SCALE_PRESETS = [80, 90, 100, 110, 120] as const;
 
-export const LYRICS_PLAIN_OPACITY_MIN = 0.5;
-export const LYRICS_PLAIN_OPACITY_MAX = 1.0;
-export const LYRICS_PLAIN_OPACITY_STEP = 0.05;
-export const LYRICS_PLAIN_OPACITY_DEFAULT = 0.9;
-export const LYRICS_PLAIN_FONT_SIZE_DEFAULT: LyricsFontSize = 'base';
-
-export const LYRICS_SYNCED_DIM_OPACITY_MIN = 0.2;
-export const LYRICS_SYNCED_DIM_OPACITY_MAX = 1.0;
-export const LYRICS_SYNCED_DIM_OPACITY_STEP = 0.05;
-export const LYRICS_SYNCED_DIM_OPACITY_DEFAULT = 0.45;
-export const LYRICS_SYNCED_FONT_SIZE_DEFAULT: LyricsFontSize = 'base';
-
-/**
- * Original synced view used past=0.25 / idle=0.45. We preserve that ratio
- * (≈ 0.5556) so when the user dims idle, past dims proportionally and stays
- * visibly fainter than idle.
- */
-export const LYRICS_SYNCED_PAST_RATIO = 0.25 / 0.45;
-
-export const LYR_SIZE_CLASS: Record<LyricsFontSize, string> = {
-  sm: 'text-sm leading-6',
-  base: 'text-base leading-7',
-  lg: 'text-lg leading-8',
-  xl: 'text-xl leading-9',
-};
+export {
+  LYRICS_PLAIN_OPACITY_MIN,
+  LYRICS_PLAIN_OPACITY_MAX,
+  LYRICS_PLAIN_OPACITY_STEP,
+  LYRICS_PLAIN_OPACITY_DEFAULT,
+  LYRICS_PLAIN_FONT_SIZE_DEFAULT,
+  LYRICS_SYNCED_DIM_OPACITY_MIN,
+  LYRICS_SYNCED_DIM_OPACITY_MAX,
+  LYRICS_SYNCED_DIM_OPACITY_STEP,
+  LYRICS_SYNCED_DIM_OPACITY_DEFAULT,
+  LYRICS_SYNCED_FONT_SIZE_DEFAULT,
+  LYRICS_SYNCED_PAST_RATIO,
+  LYR_SIZE_CLASS,
+  nextLyricsFontSize,
+} from '@/stores/useLyricsAppearanceStore';
 
 // --- Compact mode dimension presets ---
 //
@@ -84,18 +74,6 @@ export const CMP_ALBUM_CLASS: Record<CompactFontSize, string> = {
   md: 'text-[11px]',
   lg: 'text-xs',
 };
-
-const FONT_SIZE_ORDER: LyricsFontSize[] = ['sm', 'base', 'lg', 'xl'];
-
-/**
- * Active synced line uses one step larger than the user-selected base, capped
- * at xl. Mirrors the original hardcoded "base→lg active" behavior.
- */
-export function nextLyricsFontSize(size: LyricsFontSize): LyricsFontSize {
-  const idx = FONT_SIZE_ORDER.indexOf(size);
-  if (idx < 0) return 'lg';
-  return FONT_SIZE_ORDER[Math.min(idx + 1, FONT_SIZE_ORDER.length - 1)];
-}
 
 const NEW_KEY = 'shiranami.app-store';
 
@@ -157,40 +135,6 @@ function coerceUiScale(v: unknown): number {
   if (Number.isNaN(parsed)) return UI_SCALE_DEFAULT;
   return Math.round(Math.min(UI_SCALE_MAX, Math.max(UI_SCALE_MIN, parsed)));
 }
-function clampLyricsPlainOpacity(v: number): number {
-  const clamped = Math.min(LYRICS_PLAIN_OPACITY_MAX, Math.max(LYRICS_PLAIN_OPACITY_MIN, v));
-  // Round to nearest step to keep persisted values clean.
-  const steps = Math.round(clamped / LYRICS_PLAIN_OPACITY_STEP);
-  return Math.round(steps * LYRICS_PLAIN_OPACITY_STEP * 1000) / 1000;
-}
-function coerceLyricsPlainOpacity(v: unknown): number {
-  const parsed = typeof v === 'number' ? v : Number(v);
-  if (!Number.isFinite(parsed)) return LYRICS_PLAIN_OPACITY_DEFAULT;
-  return clampLyricsPlainOpacity(parsed);
-}
-function coerceLyricsPlainFontSize(v: unknown): LyricsFontSize {
-  return v === 'sm' || v === 'base' || v === 'lg' || v === 'xl'
-    ? v
-    : LYRICS_PLAIN_FONT_SIZE_DEFAULT;
-}
-function clampLyricsSyncedDimOpacity(v: number): number {
-  const clamped = Math.min(
-    LYRICS_SYNCED_DIM_OPACITY_MAX,
-    Math.max(LYRICS_SYNCED_DIM_OPACITY_MIN, v)
-  );
-  const steps = Math.round(clamped / LYRICS_SYNCED_DIM_OPACITY_STEP);
-  return Math.round(steps * LYRICS_SYNCED_DIM_OPACITY_STEP * 1000) / 1000;
-}
-function coerceLyricsSyncedDimOpacity(v: unknown): number {
-  const parsed = typeof v === 'number' ? v : Number(v);
-  if (!Number.isFinite(parsed)) return LYRICS_SYNCED_DIM_OPACITY_DEFAULT;
-  return clampLyricsSyncedDimOpacity(parsed);
-}
-function coerceLyricsSyncedFontSize(v: unknown): LyricsFontSize {
-  return v === 'sm' || v === 'base' || v === 'lg' || v === 'xl'
-    ? v
-    : LYRICS_SYNCED_FONT_SIZE_DEFAULT;
-}
 function coerceCompactSize(v: unknown): CompactSize {
   return v === 'sm' || v === 'md' || v === 'lg' ? v : COMPACT_SIZE_DEFAULT;
 }
@@ -231,10 +175,6 @@ interface PersistedAppState {
   libraryHeroCardEnabled: boolean;
   lowPerformanceMode: boolean;
   noiseOverlayEnabled: boolean;
-  lyricsPlainOpacity: number;
-  lyricsPlainFontSize: LyricsFontSize;
-  lyricsSyncedDimOpacity: number;
-  lyricsSyncedFontSize: LyricsFontSize;
   compactSize: CompactSize;
   compactFontSize: CompactFontSize;
   compactAmbientIntensity: number;
@@ -282,14 +222,6 @@ function sanitize(persisted: Partial<PersistedAppState> | undefined): Partial<Pe
     out.lowPerformanceMode = persisted.lowPerformanceMode;
   if (typeof persisted.noiseOverlayEnabled === 'boolean')
     out.noiseOverlayEnabled = persisted.noiseOverlayEnabled;
-  if (persisted.lyricsPlainOpacity !== undefined)
-    out.lyricsPlainOpacity = coerceLyricsPlainOpacity(persisted.lyricsPlainOpacity);
-  if (persisted.lyricsPlainFontSize !== undefined)
-    out.lyricsPlainFontSize = coerceLyricsPlainFontSize(persisted.lyricsPlainFontSize);
-  if (persisted.lyricsSyncedDimOpacity !== undefined)
-    out.lyricsSyncedDimOpacity = coerceLyricsSyncedDimOpacity(persisted.lyricsSyncedDimOpacity);
-  if (persisted.lyricsSyncedFontSize !== undefined)
-    out.lyricsSyncedFontSize = coerceLyricsSyncedFontSize(persisted.lyricsSyncedFontSize);
   if (persisted.compactSize !== undefined)
     out.compactSize = coerceCompactSize(persisted.compactSize);
   if (persisted.compactFontSize !== undefined)
@@ -414,10 +346,6 @@ interface AppState {
   libraryHeroCardEnabled: boolean;
   lowPerformanceMode: boolean;
   noiseOverlayEnabled: boolean;
-  lyricsPlainOpacity: number;
-  lyricsPlainFontSize: LyricsFontSize;
-  lyricsSyncedDimOpacity: number;
-  lyricsSyncedFontSize: LyricsFontSize;
   compactSize: CompactSize;
   compactFontSize: CompactFontSize;
   compactAmbientIntensity: number;
@@ -452,12 +380,6 @@ interface AppActions {
   setPlaylistGridSize: (size: AlbumGridSize) => void;
   setAlbumSortMode: (mode: AlbumSortMode) => void;
   setAlbumSortOrder: (order: AlbumSortOrder) => void;
-  setLyricsPlainOpacity: (value: number) => void;
-  setLyricsPlainFontSize: (size: LyricsFontSize) => void;
-  resetLyricsPlainAppearance: () => void;
-  setLyricsSyncedDimOpacity: (value: number) => void;
-  setLyricsSyncedFontSize: (size: LyricsFontSize) => void;
-  resetLyricsAppearance: () => void;
   setCompactSize: (size: CompactSize) => void;
   setCompactFontSize: (size: CompactFontSize) => void;
   setCompactAmbientIntensity: (value: number) => void;
@@ -491,10 +413,6 @@ export const useAppStore = create<AppState & AppActions>()(
       libraryHeroCardEnabled: true,
       lowPerformanceMode: false,
       noiseOverlayEnabled: false,
-      lyricsPlainOpacity: LYRICS_PLAIN_OPACITY_DEFAULT,
-      lyricsPlainFontSize: LYRICS_PLAIN_FONT_SIZE_DEFAULT,
-      lyricsSyncedDimOpacity: LYRICS_SYNCED_DIM_OPACITY_DEFAULT,
-      lyricsSyncedFontSize: LYRICS_SYNCED_FONT_SIZE_DEFAULT,
       compactSize: COMPACT_SIZE_DEFAULT,
       compactFontSize: COMPACT_FONT_SIZE_DEFAULT,
       compactAmbientIntensity: COMPACT_AMBIENT_INTENSITY_DEFAULT,
@@ -632,32 +550,6 @@ export const useAppStore = create<AppState & AppActions>()(
         // Scroll position is meaningless once album order changes.
         useViewStore.setState({ albumGridScrollTop: 0 });
       },
-      setLyricsPlainOpacity: value => {
-        set({ lyricsPlainOpacity: coerceLyricsPlainOpacity(value) });
-      },
-      setLyricsPlainFontSize: size => {
-        set({ lyricsPlainFontSize: coerceLyricsPlainFontSize(size) });
-      },
-      resetLyricsPlainAppearance: () => {
-        set({
-          lyricsPlainOpacity: LYRICS_PLAIN_OPACITY_DEFAULT,
-          lyricsPlainFontSize: LYRICS_PLAIN_FONT_SIZE_DEFAULT,
-        });
-      },
-      setLyricsSyncedDimOpacity: value => {
-        set({ lyricsSyncedDimOpacity: coerceLyricsSyncedDimOpacity(value) });
-      },
-      setLyricsSyncedFontSize: size => {
-        set({ lyricsSyncedFontSize: coerceLyricsSyncedFontSize(size) });
-      },
-      resetLyricsAppearance: () => {
-        set({
-          lyricsPlainOpacity: LYRICS_PLAIN_OPACITY_DEFAULT,
-          lyricsPlainFontSize: LYRICS_PLAIN_FONT_SIZE_DEFAULT,
-          lyricsSyncedDimOpacity: LYRICS_SYNCED_DIM_OPACITY_DEFAULT,
-          lyricsSyncedFontSize: LYRICS_SYNCED_FONT_SIZE_DEFAULT,
-        });
-      },
       setCompactSize: size => {
         const next = coerceCompactSize(size);
         set({ compactSize: next });
@@ -719,10 +611,6 @@ export const useAppStore = create<AppState & AppActions>()(
         libraryHeroCardEnabled: s.libraryHeroCardEnabled,
         lowPerformanceMode: s.lowPerformanceMode,
         noiseOverlayEnabled: s.noiseOverlayEnabled,
-        lyricsPlainOpacity: s.lyricsPlainOpacity,
-        lyricsPlainFontSize: s.lyricsPlainFontSize,
-        lyricsSyncedDimOpacity: s.lyricsSyncedDimOpacity,
-        lyricsSyncedFontSize: s.lyricsSyncedFontSize,
         compactSize: s.compactSize,
         compactFontSize: s.compactFontSize,
         compactAmbientIntensity: s.compactAmbientIntensity,
